+++ changelog.md
@@ -7,7 +7,8 @@ Upcoming release
 ## New features
 - New import formats
     - SPSS .zsav and .por files
-    - SAS .sas7bdat and .xpt files 
+    - SAS .sas7bdat and .xpt files
+    - STATA .dta files
 - New localizations
     - Greek (Zoé Vulgarasz)
     - Spanish (Borbála Zulauf)